requirements.txt
torch
torchvision
segmentation-models-pytorch
flask
flask-cors
h5py
numpy
matplotlib
Pillow
scipy
